squash some -Wsign-conversion warnings

+++ cmd/tools/dijkstra.c
@@ -263,8 +263,8 @@ static void init(int argc, char *argv[])
 	fprintf(stderr, "%s: no node specified\n", CmdName);
 	usage(1);
     }
-    Files = calloc(sizeof(char *), argc / 2 + 2);
-    Nodes = calloc(sizeof(char *), argc / 2 + 2);
+    Files = calloc(sizeof(char *), (size_t)argc / 2 + 2);
+    Nodes = calloc(sizeof(char *), (size_t)argc / 2 + 2);
     for (j = i = 0; i < argc; i++) {
 	Nodes[j] = argv[i++];
 	Files[j] = (argv[i] ? argv[i] : "-");
